Moves the ssh tunneling/forwarding into its own file.
